Updated match generation
The unique array list of matches is now stored in a descending way based on player availability score. The method match generation adds possible match times to a player object which will be used to then generate matches into the schedule hash map.

Working on generating the matches into the schedule hash map