Updated match generation
Can now add players into a date based on the day of the week their availability was shown to be free.

Still working on limiting the amount of matches that occur at a certain time to be based on the amount of tables open.

Also have to ensure the times either before or after have at minimum a 30 minute period in order to ensure that times do not interfere with each other.