Updated match generation
Changed enteredInfo.java file to be admInInfo.java as I believe it fits better.

Additionally, now checking to ensure that during a matches time from before and after the match is completed. If during them there are matches scheduled and it exceeds the number of tables open then that match cannot be added. Since all tables are taken for other matches or would interfere.

Working on ensuring a person cannot play multiple matches on a given day at the same time. Have to either move the time or skip the match if it cannot be moved