Updated match generation
Can now generate matches based on thirty minute intervals. The program was changed to be 30 minute intervals rather than being time flexible.

Matches can be generated but the issue is having the same player with a match scheduled twice at 2:00 for example. This overlap is still in the process of being fixed.